% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2Graph.R
\name{raster2Graph}
\alias{raster2Graph}
\title{Create a graph from an raster according the connection between cells}
\usage{
raster2Graph(r, res = NULL, silent = TRUE)
}
\arguments{
\item{r}{an object of raster. The value of each cell of the raster is the ‘smoothness’ to indicated how easy the cell connecting with neighbor cells.}

\item{res}{Numeric. Resample the input raster to given resolution and use the resampled raster to build graph. Set this to NULL if using the original resolution of of the input raster.}
}
\value{
a list with an graph and the resampled raster. The graph is igraph object, with cells as node and connections as weight
}
\description{
Create a graph from an raster according the connection between cells
}
\examples{
# read in habitat suitability data of wolf in Europe
library(raster)
hsi.file = system.file("extdata", "wolf3_int.tif", package = "habCluster")
wolf = raster(hsi.file)

# build graph from raster
g = raster2Graph(wolf, 40000)
}
